--- jnosql-cassandra-driver/src/test/resources/META-INF/services/jakarta.nosql.tck.communication.driver.column.ColumnFamilyManagerSupplier
@@ -1,1 +1,0 @@
-org.eclipse.jnosql.communication.cassandra.column.CassandraColumnFamilyManagerSupplier
