+++ jnosql-cassandra-driver/src/test/resources/META-INF/services/jakarta.nosql.tck.communication.driver.column.ColumnManagerSupplier
@@ -1,0 +1,1 @@
+org.eclipse.jnosql.communication.cassandra.column.CassandraColumnManagerSupplier
