+++ jnosql-couchdb-driver/src/test/java/org/eclipse/jnosql/communication/couchdb/document/DefaultCouchDBDocumentManagerTest.java
@@ -48,7 +48,7 @@ import static org.junit.jupiter.api.Assertions.assertNotNull;
 import static org.junit.jupiter.api.Assertions.assertThrows;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-class DefaultCouchDBDocumentCollectionManagerTest {
+class DefaultCouchDBDocumentManagerTest {
 
     public static final String COLLECTION_NAME = "person";
