+++ jnosql-elasticsearch-driver/src/test/resources/META-INF/services/jakarta.nosql.tck.communication.driver.document.DocumentManagerSupplier
@@ -1,0 +1,1 @@
+org.eclipse.jnosql.communication.elasticsearch.document.ElasticsearchDocumentManagerSupplier
