--- jnosql-solr-driver/src/test/resources/META-INF/services/jakarta.nosql.tck.communication.driver.document.DocumentCollectionManagerSupplier
@@ -1,1 +1,0 @@
-org.eclipse.jnosql.communication.solr.document.SolrDocumentCollectionManagerSupplier
